[FSSDK-10950] restructure odp directories
