+++ lib/index.browser.tests.js
@@ -26,11 +26,11 @@ import configValidator from './utils/config_validator';
 import OptimizelyUserContext from './optimizely_user_context';
 
 import { LOG_MESSAGES, ODP_EVENT_ACTION } from './utils/enums';
-import { BrowserOdpManager } from './plugins/odp_manager/index.browser';
-import { OdpConfig } from './core/odp/odp_config';
-import { BrowserOdpEventManager } from './plugins/odp/event_manager/index.browser';
-import { BrowserOdpEventApiManager } from './plugins/odp/event_api_manager/index.browser';
-import { OdpEvent } from './core/odp/odp_event';
+import { BrowserOdpManager } from './odp/odp_manager.browser';
+import { OdpConfig } from './odp/odp_config';
+import { BrowserOdpEventManager } from './odp/event_manager/event_manager.browser';
+import { BrowserOdpEventApiManager } from './odp/event_manager/event_api_manager.browser';
+import { OdpEvent } from './odp/event_manager/odp_event';
 import { getMockProjectConfigManager } from './tests/mock/mock_project_config_manager';
 import { createProjectConfig } from './project_config/project_config';
