+++ lib/index.browser.ts
@@ -26,8 +26,8 @@ import { createNotificationCenter } from './notification_center';
 import { OptimizelyDecideOption, Client, Config, OptimizelyOptions } from './shared_types';
 import { BrowserOdpManager } from './odp/odp_manager.browser';
 import Optimizely from './optimizely';
-import { IUserAgentParser } from './core/odp/user_agent_parser';
-import { getUserAgentParser } from './odp/ua_parser.browser';
+import { IUserAgentParser } from './odp/ua_parser/user_agent_parser';
+import { getUserAgentParser } from './odp/ua_parser/ua_parser.browser';
 import * as commonExports from './common_exports';
 import { PollingConfigManagerConfig } from './project_config/config_manager_factory';
 import { createPollingProjectConfigManager } from './project_config/config_manager_factory.browser';
