+++ lib/odp/segment_manager/optimizely_segment_option.ts
