+++ tests/odpEventManager.spec.ts
@@ -16,17 +16,17 @@
 import { describe, beforeEach, afterEach, beforeAll, it, vi, expect } from 'vitest';
 
 import { ODP_EVENT_ACTION, ODP_DEFAULT_EVENT_TYPE, ERROR_MESSAGES } from '../lib/utils/enums';
-import { OdpConfig } from '../lib/core/odp/odp_config';
-import { Status } from '../lib/core/odp/odp_event_manager';
-import { BrowserOdpEventManager } from "../lib/plugins/odp/event_manager/index.browser";
-import { NodeOdpEventManager } from '../lib/plugins/odp/event_manager/index.node';
-import { OdpEventManager } from '../lib/core/odp/odp_event_manager';
+import { OdpConfig } from '../lib/odp/odp_config';
+import { Status } from '../lib/odp/event_manager/odp_event_manager';
+import { BrowserOdpEventManager } from '../lib/odp/event_manager/event_manager.browser';
+import { NodeOdpEventManager } from '../lib/odp/event_manager/event_manager.node';
+import { OdpEventManager } from '../lib/odp/event_manager/odp_event_manager';
 import { anything, capture, instance, mock, resetCalls, spy, verify, when } from 'ts-mockito';
-import { IOdpEventApiManager } from '../lib/core/odp/odp_event_api_manager';
+import { IOdpEventApiManager } from '../lib/odp/event_manager/odp_event_api_manager';
 import { LogHandler, LogLevel } from '../lib/modules/logging';
-import { OdpEvent } from '../lib/core/odp/odp_event';
-import { IUserAgentParser } from '../lib/core/odp/user_agent_parser';
-import { UserAgentInfo } from '../lib/core/odp/user_agent_info';
+import { OdpEvent } from '../lib/odp/event_manager/odp_event';
+import { IUserAgentParser } from '../lib/odp/ua_parser/user_agent_parser';
+import { UserAgentInfo } from '../lib/odp/ua_parser/user_agent_info';
 import { resolve } from 'path';
 import { advanceTimersByTime } from './testUtils';
